// src/service/file/schemas/studentEvaluationDomainResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { StudentEvaluationDomainResultCourse } from './studentEvaluationDomainResultCourse';

export interface StudentEvaluationDomainResult {
  /** 영역의 코드 */
  code: string;
  /** 지표가 해당하는 교육 과정 */
  course: StudentEvaluationDomainResultCourse;
  /** 정렬 순서. 0 부터 */
  domainOrder: number;
  /** 영역 이름 */
  title: string;
  /**
   * 영역 설명
   * @nullable
   */
  description?: string | null;
  /** 생성일 */
  createdAt: string;
  /** 수정일 */
  modifiedAt: string;
}